Add bidding purchase sale tracking for solanart
As part of the new feature, we have to update the logic that detects
buyer address to check the token destination address instead of blindly
using the signer as the buyer. It's proven that in the case of a bidding
purchase, it's inaccurate to assume the signer is the buyer.